python_api/requirements.txt
charset-normalizer==2.1.1
idna==3.4
numpy==1.23.5
Pillow==9.3.0
requests==2.28.1
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
typing_extensions==4.4.0
urllib3==1.26.13
opencv-python==4.6.0.66
